backend/requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
Mako==1.3.5
MarkupSafe==2.1.5
pydantic==2.9.2
pydantic-core==2.23.4
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
typing-extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
zipp==3.20.2
